requirements.txt
google-generativeai
edge-tts
moviepy<2.0.0
requests
python-dotenv
google-auth-oauthlib
google-api-python-client
tiktok-uploader
selenium
webdriver-manager
pillow
numpy